backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "RightImpact-backend"
version = "0.1.0"
description = "Backend for RightImpact"
requires-python = ">=3.7"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.0",
    "pydantic>=1.8.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "python-multipart>=0.0.5",
    "python-dotenv>=0.19.0",
    "alembic>=1.7.0",
    "bcrypt>=3.2.0"
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests"]
